user time setting

+++ app/src/main/java/com/softsquared/template/src/task/models/SocarzoneInfo.java
@@ -3,13 +3,28 @@ package com.softsquared.template.src.task.models;
 import com.google.gson.annotations.SerializedName;
 
 public class SocarzoneInfo {
-    @SerializedName("socarzoneNo") int socarzoneNo;
-    @SerializedName("carCount") int carCount;
-    @SerializedName("latitude") float latitude;
-    @SerializedName("longitude") float longitude;
+    @SerializedName("socarzoneNo")
+    int socarzoneNo;
+    @SerializedName("carCount")
+    int carCount;
+    @SerializedName("latitude")
+    double latitude;
+    @SerializedName("longitude")
+    double longitude;
 
-    public int getSocarzoneNo() { return socarzoneNo; }
-    public int getCarCount() { return carCount; }
-    public float getLatitude() { return latitude; }
-    public float getLongitude() { return longitude; }
+    public int getSocarzoneNo() {
+        return socarzoneNo;
+    }
+
+    public int getCarCount() {
+        return carCount;
+    }
+
+    public double getLatitude() {
+        return latitude;
+    }
+
+    public double getLongitude() {
+        return longitude;
+    }
 }
